Allow retrieval of data older than 'retention' if no new data is added
.wsp file format is a circular buffer and is defined such that it doesn't
support sparse data storage. If the file capacity is 1 day, it will store
at most 1 day of data even if for 11 hours no data was received. For most
part this is a standard behavior and is mostly acceptable.

However, what if we write 3 hours of data and attempt to view that data
a week later?  The data is still there but current whisper implementation
continues wrapping because it always keys off 'now' so the result is that
while we have the data, we can't actually get it out.

This change introduces 'maxTime' file header field which is updated with
latest timestamp written to the file. Then (maxTime - retention, maxTime)
is used instead of (now - retention, now) as the data bounds.

The decision to tweak whisper rather than switching to ceres was taken
based on somewhat hazy future of ceres and megacarbon.

The good of this approach which is not possible with stock graphite
implementation:
1. We can create a small log (few days), run a test and keep the data
forever as long as no new data is loaded.
2. We can take externally captured data and feed it into graphite with
original timestamps and view it.

The bad:
1. .wsp file binary format created by this patch is not compatible with
files created by original graphite as header now has extra 4 bytes of data
2. I/O overhead is slightly increased as every file update now writes out
actual data as it did before, but then there's an additional write operation
of 4 bytes to update the header.
3. Header caching option (which isn't enabled by default) had to be
permanently disabled since header is no longer static as it was before. So
everytime we read the file we need to see if maxTime was changed.